restructure/modularize code
